Extract TeradataJdbcSelectTask.
This class will be later used in the new class to create a task to dump
Teradata utility logs.
