+++ dumper/app/src/main/java/com/google/edwmigration/dumper/application/dumper/connector/teradata/AbstractTeradataConnector.java
@@ -16,8 +16,6 @@
  */
 package com.google.edwmigration.dumper.application.dumper.connector.teradata;
 
-import com.google.common.base.Preconditions;
-import com.google.common.io.ByteSink;
 import com.google.edwmigration.dumper.application.dumper.ConnectorArguments;
 import com.google.edwmigration.dumper.application.dumper.annotations.RespectsArgumentDriverClass;
 import com.google.edwmigration.dumper.application.dumper.annotations.RespectsArgumentDriverRequired;
@@ -29,23 +27,9 @@ import com.google.edwmigration.dumper.application.dumper.annotations.RespectsInp
 import com.google.edwmigration.dumper.application.dumper.connector.AbstractJdbcConnector;
 import com.google.edwmigration.dumper.application.dumper.handle.Handle;
 import com.google.edwmigration.dumper.application.dumper.handle.JdbcHandle;
-import com.google.edwmigration.dumper.application.dumper.task.JdbcSelectTask;
-import com.google.edwmigration.dumper.application.dumper.task.TaskCategory;
-import com.google.edwmigration.dumper.application.dumper.task.TaskRunContext;
-import java.sql.Connection;
 import java.sql.Driver;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.util.List;
 import javax.annotation.Nonnull;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.dao.support.DataAccessUtils;
-import org.springframework.jdbc.core.ResultSetExtractor;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.RowMapperResultSetExtractor;
-import org.springframework.jdbc.core.SingleColumnRowMapper;
 
 /**
  * Note that connecting to any Teradata database requires their JDBC driver (not included here for
@@ -72,65 +56,7 @@ import org.springframework.jdbc.core.SingleColumnRowMapper;
 @RespectsArgumentDriverClass
 public abstract class AbstractTeradataConnector extends AbstractJdbcConnector {
 
-  @SuppressWarnings("UnusedVariable")
-  private static final Logger LOG = LoggerFactory.getLogger(AbstractTeradataConnector.class);
-
   public static final int OPT_PORT_DEFAULT = 1025;
-
-  protected static class TeradataJdbcSelectTask extends JdbcSelectTask {
-
-    private final String sqlCount;
-    private final TaskCategory category;
-
-    public TeradataJdbcSelectTask(String targetPath, TaskCategory category, String sqlTemplate) {
-      super(targetPath, String.format(sqlTemplate, "*"));
-      this.sqlCount = sqlTemplate.contains("%s") ? String.format(sqlTemplate, "count(*)") : null;
-      this.category = Preconditions.checkNotNull(category);
-    }
-
-    @Override
-    public TaskCategory getCategory() {
-      return category;
-    }
-
-    // This works to execute the count SQL on the same connection as the data SQL,
-    // because it's called from doInConnection, when we already have one connection open.
-    @Nonnull
-    private ResultSetExtractor<Void> newCountedResultSetExtractor(
-        @Nonnull ByteSink sink, @Nonnull Connection connection) throws SQLException {
-      long count = -1;
-      if (sqlCount != null) {
-        // It's a lot of infrastructure, but we don't have to write it.
-        RowMapper<Long> rowMapper = new SingleColumnRowMapper<>(Long.class);
-        ResultSetExtractor<List<Long>> resultSetExtractor =
-            new RowMapperResultSetExtractor<>(rowMapper);
-        List<Long> results = doSelect(connection, resultSetExtractor, sqlCount);
-        Long result = DataAccessUtils.nullableSingleResult(results);
-        if (result != null) count = result;
-      }
-      return newCsvResultSetExtractor(sink, count);
-    }
-
-    @Override
-    protected Void doInConnection(
-        TaskRunContext context, JdbcHandle jdbcHandle, ByteSink sink, Connection connection)
-        throws SQLException {
-      try {
-        connection.setAutoCommit(false);
-        PreparedStatement statement =
-            connection.prepareStatement(
-                "SET QUERY_BAND='ApplicationName=compilerworks;' FOR SESSION");
-        statement.execute();
-      } catch (SQLException e) {
-        // We might still be in postgresql or sqlite.
-        LOG.warn("Failed to set QUERY_BAND: " + e);
-        // This puts the transaction in an aborted state unless we rollback here.
-        connection.rollback();
-      }
-      ResultSetExtractor<Void> rse = newCountedResultSetExtractor(sink, connection);
-      return doSelect(connection, rse, getSql());
-    }
-  }
 
   /* pp */ AbstractTeradataConnector(@Nonnull String name) {
     super(name);
